requirements.txt
aiohttp==3.13.2
SQLAlchemy==2.0.35
asyncpg==0.29.0
pydantic==2.10.0
python-dotenv==1.0.1